add authorizaton ownership to tasks

+++ src/tasks/tasks.service.ts
@@ -1,5 +1,6 @@
 import { Injectable, NotFoundException } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
+import { User } from 'src/auth/user.entity';
 import { CreateTaskDto } from './dto/create-task.dto';
 import { GetTasksFilterDto } from './dto/get-tasks-filter.dto';
 import { TASK_STATUS } from './task.constant';
@@ -13,30 +14,30 @@ export class TasksService {
 
     }
 
-    async getTasks(filterDto: GetTasksFilterDto): Promise<Task[]>{
-        return await this.taskRepository.getTasks(filterDto);
+    async getTasks(filterDto: GetTasksFilterDto, user: User): Promise<Task[]>{
+        return await this.taskRepository.getTasks(filterDto, user);
     }
 
-    async getTaskById(id: number): Promise<Task>{
-        const found = await this.taskRepository.findOne(id);
+    async getTaskById(id: number, user: User): Promise<Task>{
+        const found = await this.taskRepository.findOne({where: {id, userId: user.id}});
         if(!found)
             throw new NotFoundException(`Task with id "${id}" not found`);
         return found;
     }
 
-    async createTask(createTaskDto: CreateTaskDto): Promise<Task>{
-        return this.taskRepository.createTask(createTaskDto);
+    async createTask(createTaskDto: CreateTaskDto, user: User): Promise<Task>{
+        return this.taskRepository.createTask(createTaskDto, user);
     }
 
-    async updateTaskStatus(id: number, status: TASK_STATUS): Promise<Task>{
-        const task = await this.getTaskById(id);
+    async updateTaskStatus(id: number, status: TASK_STATUS, user: User): Promise<Task>{
+        const task = await this.getTaskById(id, user);
         task.status = status;
         await task.save();
         return task;
     }
 
-    async deleteTask(id: number): Promise<void>{
-       const result = await this.taskRepository.delete(id);
+    async deleteTask(id: number, user: User): Promise<void>{
+       const result = await this.taskRepository.delete({id, userId: user.id});
        
        if(result.affected === 0)
         throw new NotFoundException(`Task with id "${id}" not found`);
